+++ src/App.js
@@ -4,7 +4,7 @@ import "./components/defaultCss/style.css";
 import Menu from "./components/menu/menu";
 import Header from "./components/header/header";
 import Sobre from "./components/sobre/sobre";
-import Portfolio from "./components/portfolio/portfolio";
+import Portfolio from "./components/portfolio/";
 import Contato from "./components/contato/contato";
 import Footer from "./components/footer/footer";
