+++ src/components/portfolio/repositoriesList/tools/isFork.js
@@ -1,9 +1,8 @@
 import { BiGitRepoForked } from "react-icons/bi";
 
-export default class isFork {
+export default class IsFork {
   static check(yesOrNot) {
     if (yesOrNot === false) return;
-
-    return <BiGitRepoForked />;
+    else return <BiGitRepoForked />;
   }
 }
